// speedyGo-ui/src/app/services/models/event-promotion.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Product } from '../models/product';
import { User } from '../models/user';
export interface EventPromotion {
  createdDate?: string;
  customerEvent_Promotion?: Array<User>;
  description?: string;
  discount?: number;
  endDate?: string;
  id?: number;
  imageUrl?: string;
  lastModifiedDate?: string;
  partnerEvent_Promotion?: Array<User>;
  products?: Product;
  startDate?: string;
  status?: 'ACTIVE' | 'EXPIRED' | 'UPCOMING';
  typeEV_ep?: 'EVENT' | 'PROMOTION';
  productId?:number;

}